+++ src/schema2ts/convertSwagger.ts
@@ -11,7 +11,7 @@ const filterStringByMapping = async (text: string | string[]) =>
 
 export const swaggerSchemaBasicTypes = ['string', 'boolean', 'number', 'integer'];
 
-export const convertAPIV2Schema2JSONSchema = async (swaggerSchema: OpenAPIV2.SchemaObject, isArrayItems = false): Promise<JSONSchema> => {
+export const convertAPIV2Schema2JSONSchema = async (swaggerSchema: OpenAPIV2.SchemaObject): Promise<JSONSchema> => {
   const { $ref: $schemaRef, type: schemaType = '', title: schemaTitle, description, ...otherProps } = swaggerSchema;
   const mergeCommonJSONSchema = (additionalSchema: JSONSchema): JSONSchema => ({ title: schemaTitle, description, ...additionalSchema });
 
@@ -19,11 +19,6 @@ export const convertAPIV2Schema2JSONSchema = async (swaggerSchema: OpenAPIV2.Sch
   if ($schemaRef) {
     const refClassName = match$RefClassName($schemaRef);
     const currentClassName = await filterStringByMapping(refClassName);
-
-    // 引用类型数组直接定义 tsType
-    if (isArrayItems) {
-      return { tsType: currentClassName };
-    }
 
     return {
       $ref: `#/definitions/${currentClassName}`,
@@ -56,7 +51,7 @@ export const convertAPIV2Schema2JSONSchema = async (swaggerSchema: OpenAPIV2.Sch
     const { items: arrayItems } = swaggerSchema;
     return mergeCommonJSONSchema({
       type: 'array',
-      items: await convertAPIV2ToJSONSchema(arrayItems ?? {}, null, true),
+      items: await convertAPIV2ToJSONSchema(arrayItems ?? {}, null),
     });
   }
 
@@ -89,9 +84,8 @@ export const convertAPIV2Schema2JSONSchema = async (swaggerSchema: OpenAPIV2.Sch
 export const convertAPIV2ToJSONSchema = async (
   swaggerSchema: OpenAPIV2.SchemaObject,
   V2Document?: OpenAPIV2.Document | null,
-  isArrayItems = false,
 ): Promise<JSONSchema> => {
-  const JSONSchema = await convertAPIV2Schema2JSONSchema(swaggerSchema, isArrayItems);
+  const JSONSchema = await convertAPIV2Schema2JSONSchema(swaggerSchema);
   JSONSchema.definitions = await convertAPIV2Definitions(
     (swaggerSchema.definitions ? (swaggerSchema.definitions as OpenAPIV2.DefinitionsObject) : V2Document?.definitions) ?? {},
   );