+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "tswagger" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Missing types caused by tsType.
+
 ## [1.0.6] - 2023-12-12
 
 ### Fixed