+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "tswagger" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Generate other interfaces failed in the same group.
+
 ## [1.0.7] - 2023-12-12
 
 ### Fixed